refactor(ZNTA-1735): rename component module
