+++ server/zanata-frontend/src/frontend/app/editor/components/GlossaryTermModal/component.js
